fix: check user existence

+++ src/main/java/controllers/UserController.java
@@ -22,6 +22,9 @@ public class UserController extends Controller {
 
     public void deleteUser(String username) {
         User user = dao.getUser(username);
+        if(user == null) {
+            throw new IllegalArgumentException();
+        }
         dao.delete(user);
     }
